refact: cart popup and search bar component created
